Further clean up TropoDB
