+++ implementation/rocksdb/db/zns_impl/db_impl_zns.cc
@@ -44,15 +44,8 @@
 
 namespace ROCKSDB_NAMESPACE {
 
+// TODO: we do not use this? Remove?
 const int kNumNonTableCacheFiles = 10;
-
-struct DBImplZNS::Writer {
-  explicit Writer(port::Mutex* mu) : batch(nullptr), done(false), cv(mu) {}
-  Status status;
-  WriteBatch* batch;
-  bool done;
-  port::CondVar cv;
-};
 
 DBImplZNS::DBImplZNS(const DBOptions& options, const std::string& dbname,
                      const bool seq_per_batch, const bool batch_per_txn,
@@ -68,6 +61,9 @@ DBImplZNS::DBImplZNS(const DBOptions& options, const std::string& dbname,
       manifest_(nullptr),
       table_cache_(nullptr),
       versions_(nullptr),
+      // Thread count (1 HIGH for each flush thread, 1~3 HIGH for each L0 thread and 1~3 LOW for each LN thread)
+      low_level_threads_((1 + ZnsConfig::compaction_allow_prefetching + ZnsConfig::compaction_allow_deferring_writes)),
+      high_level_threads_(ZnsConfig::lower_concurrency + ZnsConfig::lower_concurrency * (1 + ZnsConfig::compaction_allow_prefetching + ZnsConfig::compaction_allow_deferring_writes)),
       // State
       bg_work_l0_finished_signal_(&mutex_),
       bg_work_finished_signal_(&mutex_),
@@ -80,6 +76,7 @@ DBImplZNS::DBImplZNS(const DBOptions& options, const std::string& dbname,
       // diag
       clock_(SystemClock::Default().get()) {
 	SetTropoDBLogLevel(ZnsConfig::default_log_level);
+  // The following variables are set to safeguards
   for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
     wal_man_[i] = nullptr;
     wal_[i] = nullptr;
@@ -89,54 +86,60 @@ DBImplZNS::DBImplZNS(const DBOptions& options, const std::string& dbname,
     tmp_batch_[i] = new WriteBatch;
     wal_reserved_[i] = 0;
   }
-  for (uint8_t i = 0; i < ZnsConfig::level_count - 1; i++) {
-    compactions_[i] = 0;
-  }
-  env_->SetBackgroundThreads(2, ROCKSDB_NAMESPACE::Env::Priority::HIGH);
-  env_->SetBackgroundThreads(5, ROCKSDB_NAMESPACE::Env::Priority::LOW);
+  std::fill(compactions_.begin(), compactions_.end(), 0);
+  // Setup background threads (RocksDB env will not let us make threads otherwise)
+  TROPODB_INFO("INFO: TropoDB will use a maximum of %u background threads\n", low_level_threads_ + high_level_threads_);
+  env_->SetBackgroundThreads(high_level_threads_, ROCKSDB_NAMESPACE::Env::Priority::HIGH);
+  env_->SetBackgroundThreads(low_level_threads_, ROCKSDB_NAMESPACE::Env::Priority::LOW);
 }
 
 DBImplZNS::~DBImplZNS() {
-  printf("Shutdown \n");
-  mutex_.Lock();
-  while (bg_compaction_l0_scheduled_ || bg_compaction_scheduled_ ||
-         AnyFlushScheduled()) {
-    shutdown_ = true;
-    printf("busy, wait before closing\n");
-    if (bg_compaction_l0_scheduled_) {
-      bg_work_l0_finished_signal_.Wait();
+  TROPODB_INFO("INFO: Shutting down TropoDB\n");
+  // Close all tasks
+  {
+    mutex_.Lock();
+    while (bg_compaction_l0_scheduled_ || bg_compaction_scheduled_ ||
+          AnyFlushScheduled()) {
+      shutdown_ = true;
+      printf("busy, wait before closing\n");
+      if (bg_compaction_l0_scheduled_) {
+        bg_work_l0_finished_signal_.Wait();
+      }
+      if (bg_compaction_scheduled_) {
+        bg_work_finished_signal_.Wait();
+      }
+      if (AnyFlushScheduled()) {
+        bg_flush_work_finished_signal_.Wait();
+      }
     }
-    if (bg_compaction_scheduled_) {
-      bg_work_finished_signal_.Wait();
+    for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
+      if (wal_[i] != nullptr) {
+        wal_[i]->Sync();
+      }
     }
-    if (AnyFlushScheduled()) {
-      bg_flush_work_finished_signal_.Wait();
-    }
+    mutex_.Unlock();
   }
-  for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
-    if (wal_[i] != nullptr) {
-      wal_[i]->Sync();
-    }
-  }
-  mutex_.Unlock();
-
+  TROPODB_INFO("INFO: All jobs done - ready to exit\n");
   PrintStats();
 
-  if (versions_ != nullptr) delete versions_;
-  for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
-    if (mem_[i] != nullptr) mem_[i]->Unref();
-    if (imm_[i] != nullptr) imm_[i]->Unref();
-    if (wal_[i] != nullptr) wal_[i]->Unref();
-    if (wal_man_[i] != nullptr) wal_man_[i]->Unref();
-    if (tmp_batch_[i] != nullptr) delete tmp_batch_[i];
-  }
-  if (ss_manager_ != nullptr) ss_manager_->Unref();
-  if (manifest_ != nullptr) manifest_->Unref();
-  if (table_cache_ != nullptr) delete table_cache_;
-  if (channel_factory_ != nullptr) channel_factory_->Unref();
-  if (zns_device_ != nullptr) delete zns_device_;
-  // printf("exiting \n");
-}  // namespace ROCKSDB_NAMESPACE
+ // Reaping what we have sown with ref counters
+ {
+    if (versions_ != nullptr) delete versions_;
+    for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
+      if (mem_[i] != nullptr) mem_[i]->Unref();
+      if (imm_[i] != nullptr) imm_[i]->Unref();
+      if (wal_[i] != nullptr) wal_[i]->Unref();
+      if (wal_man_[i] != nullptr) wal_man_[i]->Unref();
+      if (tmp_batch_[i] != nullptr) delete tmp_batch_[i];
+    }
+    if (ss_manager_ != nullptr) ss_manager_->Unref();
+    if (manifest_ != nullptr) manifest_->Unref();
+    if (table_cache_ != nullptr) delete table_cache_;
+    if (channel_factory_ != nullptr) channel_factory_->Unref();
+    if (zns_device_ != nullptr) delete zns_device_;
+ }
+ TROPODB_INFO("INFO: Exiting\n");
+}
 
 Status DBImplZNS::ValidateOptions(const DBOptions& db_options) {
   if (db_options.db_paths.size() > 1) {
@@ -154,11 +157,13 @@ Status DBImplZNS::OpenZNSDevice(const std::string dbname) {
   Status s;
   s = FromStatus(zns_device_->Init());
   if (!s.ok()) {
+    TROPODB_ERROR("ERROR: SPDK will not init. Are you root?\n");
     return Status::IOError("Error opening SPDK");
   }
   s = FromStatus(
       zns_device_->Open(this->name_, ZnsConfig::min_zone, ZnsConfig::max_zone));
   if (!s.ok()) {
+    TROPODB_ERROR("ERROR: SZD device does not open. Is it ZNS?\n");
     return Status::IOError("Error opening ZNS device");
   }
   channel_factory_ = new SZD::SZDChannelFactory(zns_device_->GetDeviceManager(),
@@ -174,9 +179,13 @@ Status DBImplZNS::ResetZNSDevice() {
   s = FromStatus(channel_factory_->register_channel(&channel));
   if (s.ok()) {
     s = FromStatus(channel->ResetAllZones());
+  } else {
+    TROPODB_ERROR("ERROR: Can not create I/O QPair for clearing device\n");
   }
   if (s.ok()) {
     s = FromStatus(channel_factory_->unregister_channel(channel));
+  } else {
+    TROPODB_ERROR("ERROR: Can not clear device \n");
   }
   channel_factory_->Unref();
   return s.ok() ? Status::OK() : Status::IOError("Error resetting device");
@@ -184,70 +193,96 @@ Status DBImplZNS::ResetZNSDevice() {
 
 Status DBImplZNS::InitDB(const DBOptions& options,
                          const size_t max_write_buffer_size) {
-  max_write_buffer_size_ = max_write_buffer_size;
   assert(zns_device_ != nullptr);
-  SZD::DeviceInfo device_info;
-  zns_device_->GetInfo(&device_info);
-  uint64_t zone_head = device_info.min_lba / device_info.zone_size;
-  uint64_t zone_step = 0;
-
-  std::cout << "==== Zone division ====\n";
-  std::cout << std::left << std::setw(15) << "Structure" << std::right
+  max_write_buffer_size_ = max_write_buffer_size;
+  
+  // Setup info string
+  std::ostringstream info_str;
+  info_str << "==== Zone division ====\n";
+  info_str << std::setfill('-') << std::setw(76) << "\n" << std::setfill(' ');
+  info_str << std::left << std::setw(15) << "Structure" << std::right
             << std::setw(25) << "Begin (zone nr)" << std::setw(25)
             << "End (zone nr)"
             << "\n";
-  std::cout << std::setfill('_') << std::setw(76) << "\n" << std::setfill(' ');
-  zone_step = ZnsConfig::manifest_zones;
-  manifest_ = new ZnsManifest(channel_factory_, device_info, zone_head,
-                              zone_head + zone_step);
-  manifest_->Ref();
-  std::cout << std::left << std::setw(15) << "Manifest" << std::right
-            << std::setw(25) << zone_head << std::setw(25)
-            << zone_head + zone_step << "\n";
-  zone_head += zone_step;
+  info_str << std::setfill('-') << std::setw(76) << "\n" << std::setfill(' ');
 
-  zone_step = ZnsConfig::wal_count * ZnsConfig::zones_foreach_wal;
-  zone_step /= ZnsConfig::lower_concurrency;
-  for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
-    wal_man_[i] = new ZnsWALManager<ZnsConfig::wal_manager_zone_count>(
-        channel_factory_, device_info, zone_head, zone_step + zone_head);
-    wal_man_[i]->Ref();
+  // Get device info
+  SZD::DeviceInfo device_info;
+  zns_device_->GetInfo(&device_info);  
+  uint64_t zone_head = device_info.min_lba / device_info.zone_size;
+  uint64_t zone_step = 0;
+
+  // Init manifest
+  {
+    zone_step = ZnsConfig::manifest_zones;
+    manifest_ = new ZnsManifest(channel_factory_, device_info, zone_head,
+                                zone_head + zone_step);
+    manifest_->Ref();
+    info_str << std::left << std::setw(15) << "Manifest" << std::right
+              << std::setw(25) << zone_head 
+              << std::setw(25) << zone_head + zone_step << "\n";
     zone_head += zone_step;
   }
-  zone_step = device_info.max_lba / device_info.zone_size - zone_head - 1;
-  // If only we had access to C++23.
-  ss_manager_ =
-      ZNSSSTableManager::NewZNSSTableManager(channel_factory_, device_info,
-                                             zone_head, zone_head + zone_step)
-          .value_or(nullptr);
-  if (ss_manager_ == nullptr) {
-    return Status::Corruption();
-  }
-  ss_manager_->Ref();
-  std::cout << ss_manager_->LayoutDivisionString();
-  std::cout << std::setfill('_') << std::setw(76) << "\n" << std::setfill(' ');
-  zone_head = device_info.max_lba / device_info.zone_size;
 
-  for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
-    mem_[i] =
-        new ZNSMemTable(options, internal_comparator_, max_write_buffer_size_);
-    mem_[i]->Ref();
+  // Init WALs
+  {
+    zone_step = ZnsConfig::wal_count * ZnsConfig::zones_foreach_wal;
+    zone_step /= ZnsConfig::lower_concurrency;
+    for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
+      wal_man_[i] = new ZnsWALManager<ZnsConfig::wal_manager_zone_count>(
+          channel_factory_, device_info, zone_head, zone_step + zone_head);
+      wal_man_[i]->Ref();
+      info_str << std::left << std::setw(15) << ("WALMAN-" + std::to_string(i)) << std::right 
+               << std::setw(25) << zone_head 
+               << std::setw(25) << (zone_head + zone_step) << "\n";
+      zone_head += zone_step;
+    }
   }
 
-  Options opts(options, ColumnFamilyOptions());
-  table_cache_ = new ZnsTableCache(opts, internal_comparator_, 1024 * 1024 * 4,
+  // Init SSTable manager
+  {
+    zone_step = device_info.max_lba / device_info.zone_size - zone_head - 1;
+    // If only we had access to C++23.
+    ss_manager_ =
+        ZNSSSTableManager::NewZNSSTableManager(channel_factory_, device_info,
+                                              zone_head, zone_head + zone_step)
+            .value_or(nullptr);
+    if (ss_manager_ == nullptr) {
+      TROPODB_ERROR("ERROR: Could not initialise SSTable manager\n");
+      return Status::Corruption();
+    }
+    ss_manager_->Ref();
+    info_str << ss_manager_->LayoutDivisionString();    
+    zone_head = device_info.max_lba / device_info.zone_size;
+  }
+
+  // Init Memtables, table ache and version structure
+  {
+    for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
+      mem_[i] =
+          new ZNSMemTable(options, internal_comparator_, max_write_buffer_size_);
+      mem_[i]->Ref();
+    }
+
+    Options opts(options, ColumnFamilyOptions());
+    table_cache_ = new ZnsTableCache(opts, internal_comparator_, 1024 * 1024 * 4,
                                    ss_manager_);
 
-  versions_ = new ZnsVersionSet(internal_comparator_, ss_manager_, manifest_,
+    versions_ = new ZnsVersionSet(internal_comparator_, ss_manager_, manifest_,
                                 device_info.lba_size, device_info.zone_cap,
                                 table_cache_, this->env_);
+  }
+
+  // Print info string (if enabled)
+  info_str << std::setfill('-') << std::setw(76) << "\n" << std::setfill(' ');
+  layout_string_ = info_str.str();
+  TROPODB_INFO("%s", layout_string_.data());
 
   return Status::OK();
 }
 
-Status DBImplZNS::InitWAL() {
-  mutex_.AssertHeld();
-  Status s;
+void DBImplZNS::RecoverBackgroundFlow() {
+  // Make WALs lively again
   for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
     // We must force flush all WALs if there is not enough space.
     if (!wal_man_[i]->WALAvailable()) {
@@ -265,8 +300,46 @@ Status DBImplZNS::InitWAL() {
     wal_[i] = wal_man_[i]->GetCurrentWAL(&mutex_);
     wal_[i]->Ref();
   }
+  // Reinstigate background jobs
+  for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
+    MaybeScheduleFlush(i);
+  }
   MaybeScheduleCompactionL0();
-  MaybeScheduleCompaction(true);
+  MaybeScheduleCompaction(false);
+}
+
+Status DBImplZNS::Recover() {
+  TROPODB_INFO("INFO: recovering TropoDB\n");
+  Status s;
+
+  // Recover index structure
+  s = versions_->Recover();
+  // If there is no version to be recovered, we assume there is no valid DB.
+  if (!s.ok()) {
+    return options_.create_if_missing ? ResetZNSDevice() : s;
+    // TODO: this is not enough when version is corrupt, then device WILL be
+    // reset, but metadata still points to corrupt.
+  }
+  // TODO: currently this still writes a new version... if not an identical
+  // one.
+  if (options_.error_if_exists) {
+    return Status::InvalidArgument("DB already exists");
+  }
+
+  // Recover WAL and MVCC
+  {
+    SequenceNumber old_seq;
+    for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
+      s = wal_man_[i]->Recover(mem_[i], &old_seq);
+    }
+    if (!s.ok()) { 
+      return s;
+    }
+    versions_->SetLastSequence(old_seq);
+  }
+
+  // Recover flow
+  RecoverBackgroundFlow();
   return s;
 }
 
@@ -278,37 +351,37 @@ Status DBImplZNS::Open(
   Status s;
   s = ValidateOptions(db_options);
   if (!s.ok()) {
+    TROPODB_ERROR("ERROR: Invalid options for TropoDB\n");
     return s;
   }
   // We do not support column families, so we just clear them
   handles->clear();
 
+  // Set write buffer size to an acceptable level
   size_t max_write_buffer_size = 0;
   for (auto cf : column_families) {
     max_write_buffer_size =
         std::max(max_write_buffer_size, cf.options.write_buffer_size);
   }
 
+  // Open a SZD connection
   DBImplZNS* impl = new DBImplZNS(db_options, name);
   s = impl->OpenZNSDevice("ZNSLSM");
-  if (!s.ok()) return s;
+  if (!s.ok()) { 
+    return s;
+  }
+
+  // Divide device and recover/create components
   s = impl->InitDB(db_options, max_write_buffer_size);
-  if (!s.ok()) return s;
-  // setup WAL (WAL DIR)
+  if (!s.ok()) {
+    return s;
+  }
+  // Lock to ensure we are master of TropoDB, not bg threads
   impl->mutex_.Lock();
   s = impl->Recover();
-  if (s.ok()) {
-    s = impl->InitWAL();
-  }
-  if (s.ok()) {
-    // impl->RemoveObsoleteZones();
-    for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
-      impl->MaybeScheduleFlush(i);
-    }
-    impl->MaybeScheduleCompactionL0();
-    impl->MaybeScheduleCompaction(false);
-  }
   impl->mutex_.Unlock();
+
+  // Return DB or null based on state
   if (s.ok()) {
     *dbptr = reinterpret_cast<DB*>(impl);
   } else {
@@ -327,6 +400,8 @@ bool DBImplZNS::AnyFlushScheduled() {
 }
 
 Status DBImplZNS::Close() {
+  // Wait till all jobs are done
+  TROPODB_INFO("INFO: Closing: waiting for background jobs to finish\n");
   mutex_.Lock();
   while (bg_compaction_l0_scheduled_ || bg_compaction_scheduled_ ||
          AnyFlushScheduled()) {
@@ -344,362 +419,31 @@ Status DBImplZNS::Close() {
   }
   mutex_.Unlock();
   // TODO: close device.
+  TROPODB_ERROR("ERROR: Closing: Not implemented\n");
   return Status::OK();
-}
-
-Status DBImplZNS::Recover() {
-  printf("recovering\n");
-  Status s;
-  // Recover index structure
-  s = versions_->Recover();
-  // If there is no version to be recovered, we assume there is no valid DB.
-  if (!s.ok()) {
-    return options_.create_if_missing ? ResetZNSDevice() : s;
-    // TODO: this is not enough when version is corrupt, then device WILL be
-    // reset, but metadata still points to corrupt.
-  }
-  // TODO: currently this still writes a new version... if not an identical
-  // one.
-  if (options_.error_if_exists) {
-    return Status::InvalidArgument("DB already exists");
-  }
-
-  // Recover WAL and head
-  SequenceNumber old_seq;
-  for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
-    s = wal_man_[i]->Recover(mem_[i], &old_seq);
-  }
-  if (!s.ok()) return s;
-  versions_->SetLastSequence(old_seq);
-  return s;
 }
 
 Status DBImplZNS::DestroyDB(const std::string& dbname, const Options& options) {
-  // Destroy "all" files from the DB. Since we do not use multitenancy, we
+  // Destroy "all files" from the DB. Since we do not use multitenancy, we
   // might as well reset the device.
   Status s;
   DBImplZNS* impl = new DBImplZNS(options, dbname);
+  TROPODB_INFO("INFO: Attemtping to reset entire ZNS device\n");
   s = impl->OpenZNSDevice("ZNSLSM");
-  if (!s.ok()) return s;
+  if (!s.ok()) {
+    return s;
+  }
   s = impl->InitDB(options, options.write_buffer_size);
-  if (!s.ok()) return s;
+  if (!s.ok()) {
+    return s;
+  }
   s = impl->ResetZNSDevice();
-  if (!s.ok()) return s;
-  printf("Reset device\n");
+  if (!s.ok()) {
+    return s;
+  }
+  TROPODB_INFO("INFO: Entire ZNS device has been reset\n");
   delete impl;
   return s;
-}
-
-// Default implementations of convenience methods that subclasses of DB
-// can call if they wish
-Status DBImplZNS::Put(const WriteOptions& options, const Slice& key,
-                      const Slice& value) {
-  WriteBatch batch;
-  batch.Put(key, value);
-  return Write(options, &batch);
-}
-
-Status DBImplZNS::Delete(const WriteOptions& opt, const Slice& key) {
-  WriteBatch batch;
-  batch.Delete(key);
-  return Write(opt, &batch);
-}
-
-Status DBImplZNS::MakeRoomForWrite(size_t size, uint8_t parallel_number) {
-  mutex_.AssertHeld();
-  Status s;
-  bool allow_delay = true;
-  while (true) {
-    if (!bg_error_.ok()) {
-      // Yield error
-      s = bg_error_;
-      return s;
-    }
-    if (allow_delay && versions_->NumLevelZones(0) > ZnsConfig::L0_slow_down) {
-      mutex_.Unlock();
-      // printf("SlowDown reached...\n");
-      env_->SleepForMicroseconds(1000);
-      allow_delay = false;
-      mutex_.Lock();
-    } else if (!mem_[parallel_number]->ShouldScheduleFlush() &&
-               wal_[parallel_number]->SpaceLeft(size)) {
-      // space left in memory table
-      break;
-    } else if (imm_[parallel_number] != nullptr) {
-      // flush is scheduled, wait...
-      // printf("Imm already scheduled\n");
-      bg_flush_work_finished_signal_.Wait();
-    } else if (versions_->NeedsL0Compaction()) {
-      // No more space in L0... Better to wait till compaction is done
-      printf("Force L0\n");
-      MaybeScheduleCompactionL0();
-      bg_work_l0_finished_signal_.Wait();
-    } else if (!wal_man_[parallel_number]->WALAvailable()) {
-      printf("Out of WALs\n");
-      bg_flush_work_finished_signal_.Wait();
-    } else {
-      // create new WAL
-      wal_[parallel_number]->Sync();
-      wal_[parallel_number]->Close();
-      wal_[parallel_number]->Unref();
-      s = wal_man_[parallel_number]->NewWAL(&mutex_, &wal_[parallel_number]);
-      wal_[parallel_number]->Ref();
-#ifdef WALPerfTest
-      // Drop all that was in the memtable (NOT PERSISTENT!)
-      mem_[parallel_number]->Unref();
-      mem_[parallel_number] = new ZNSMemTable(options_, internal_comparator_,
-                                              max_write_buffer_size_);
-      mem_[parallel_number]->Ref();
-      env_->Schedule(&DBImplZNS::BGFlushWork, this, rocksdb::Env::HIGH);
-#else
-      // printf("Reset WAL\n");
-      // Switch to fresh memtable
-      imm_[parallel_number] = mem_[parallel_number];
-      mem_[parallel_number] = new ZNSMemTable(options_, internal_comparator_,
-                                              max_write_buffer_size_);
-      mem_[parallel_number]->Ref();
-      MaybeScheduleFlush(parallel_number);
-      MaybeScheduleCompactionL0();
-#endif
-    }
-  }
-  return Status::OK();
-}  // namespace ROCKSDB_NAMESPACE
-
-WriteBatch* DBImplZNS::BuildBatchGroup(Writer** last_writer,
-                                       uint8_t parallel_number) {
-  mutex_.AssertHeld();
-  assert(!writers_[parallel_number].empty());
-  Writer* first = writers_[parallel_number].front();
-  WriteBatch* result = first->batch;
-  assert(result != nullptr);
-
-  size_t size = WriteBatchInternal::ByteSize(first->batch);
-
-  // Allow the group to grow up to a maximum size, but if the
-  // original write is small, limit the growth so we do not slow
-  // down the small write too much.
-  size_t max_size =
-      max_write_buffer_size_ - 512;  // TODO: if max_write_buffer_size_ < 512
-                                     // then all hell breaks loose!
-  // if (size <= (128 << 10)) {
-  //   max_size = size + (128 << 10);
-  // }
-  if (max_size > wal_[parallel_number]->SpaceAvailable()) {
-    max_size = wal_[parallel_number]->SpaceAvailable();
-  }
-
-  *last_writer = first;
-  std::deque<Writer*>::iterator iter = writers_[parallel_number].begin();
-  ++iter;  // Advance past "first"
-  for (; iter != writers_[parallel_number].end(); ++iter) {
-    Writer* w = *iter;
-    if (w->batch != nullptr) {
-      size += WriteBatchInternal::ByteSize(w->batch);
-      if (size > max_size) {
-        // Do not make batch too big
-        break;
-      }
-
-      // Append to *result
-      if (result == first->batch) {
-        // Switch to temporary batch instead of disturbing caller's batch
-        result = tmp_batch_[parallel_number];
-        assert(WriteBatchInternal::Count(result) == 0);
-        WriteBatchInternal::Append(result, first->batch);
-      }
-      WriteBatchInternal::Append(result, w->batch);
-    }
-    *last_writer = w;
-  }
-  return result;
-}
-
-Status DBImplZNS::Write(const WriteOptions& options, WriteBatch* updates) {
-  Status s;
-
-  Writer w(&mutex_);
-  w.batch = updates;
-  w.done = false;
-  MutexLock l(&mutex_);
-
-  uint8_t striped_index = writer_striper_;
-  writer_striper_ = writer_striper_ + 1 == ZnsConfig::lower_concurrency
-                        ? 0
-                        : writer_striper_ + 1;
-  writers_[striped_index].push_back(&w);
-  while (!w.done && &w != writers_[striped_index].front()) {
-    w.cv.Wait();
-  }
-  if (w.done) {
-    return w.status;
-  }
-
-  s = MakeRoomForWrite(updates == nullptr
-                           ? 0
-                           : WriteBatchInternal::Contents(updates).size() +
-                                 wal_reserved_[striped_index],
-                       striped_index);
-  uint64_t last_sequence = versions_->LastSequence();
-  Writer* last_writer = &w;
-  // Write to what is needed
-  if (s.ok() && updates != nullptr) {
-    WriteBatch* write_batch = BuildBatchGroup(&last_writer, striped_index);
-    WriteBatchInternal::SetSequence(write_batch, last_sequence + 1);
-    last_sequence += WriteBatchInternal::Count(write_batch);
-    {
-      wal_reserved_[striped_index] =
-          WriteBatchInternal::Contents(write_batch).size();
-      wal_[striped_index]->Ref();
-      mutex_.Unlock();
-      // buffering does not really make sense at the moment.
-      // later we might decide to implement "FSync" or "ZoneSync", then it
-      // should be reinstagated.
-      // printf("writing %lu/%lu  %lu\n", mem_->GetInternalSize(),
-      //        max_write_buffer_size_,
-      //        WriteBatchInternal::Contents(write_batch).size());
-      if (!options.sync) {
-        s = wal_[striped_index]->Append(
-            WriteBatchInternal::Contents(write_batch), last_sequence + 1);
-      } else {
-        s = wal_[striped_index]->Append(
-            WriteBatchInternal::Contents(write_batch), last_sequence + 1);
-        s = wal_[striped_index]->Sync();
-      }
-      // write to memtable
-      assert(mem_[striped_index] != nullptr);
-      if (s.ok()) {
-        s = mem_[striped_index]->Write(options, write_batch);
-        if (!s.ok()) {
-          printf("Error writing to memtable %s\n", s.getState());
-        }
-      }
-      mutex_.Lock();
-      wal_[striped_index]->Unref();
-    }
-    if (write_batch == tmp_batch_[striped_index]) {
-      tmp_batch_[striped_index]->Clear();
-    }
-    versions_->SetLastSequence(last_sequence);
-  }
-
-  while (true) {
-    Writer* ready = writers_[striped_index].front();
-    writers_[striped_index].pop_front();
-    if (ready != &w) {
-      ready->status = s;
-      ready->done = true;
-      ready->cv.Signal();
-    }
-    if (ready == last_writer) break;
-  }
-
-  if (!writers_[striped_index].empty()) {
-    writers_[striped_index].front()->cv.Signal();
-  }
-
-  return s;
-}
-
-Status DBImplZNS::Get(const ReadOptions& options, const Slice& key,
-                      std::string* value) {
-  MutexLock l(&mutex_);
-  Status s;
-  value->clear();
-  // This is absolutely necessary for locking logic because private pointers
-  // can be changed in background work.
-  std::vector<ZNSMemTable*> mem;
-  std::vector<ZNSMemTable*> imm;
-  for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
-    mem.push_back(mem_[i]);
-    imm.push_back(imm_[i]);
-    mem[i]->Ref();
-    if (imm[i] != nullptr) imm[i]->Ref();
-  }
-
-  ZnsVersion* current = versions_->current();
-
-  current->Ref();
-  LookupKey lkey(key, versions_->LastSequence());
-  {
-    mutex_.Unlock();
-    SequenceNumber seq;
-    SequenceNumber seq_pot;
-    bool found = false;
-    std::string tmp;
-    for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
-      if (mem[i]->Get(options, lkey, &tmp, &s, &seq_pot)) {
-        if (!found) {
-          found = true;
-          seq = seq_pot;
-          *value = tmp;
-        } else if (seq_pot > seq) {
-          seq = seq_pot;
-          *value = tmp;
-        }
-      } else if (imm[i] != nullptr &&
-                 imm[i]->Get(options, lkey, &tmp, &s, &seq_pot)) {
-        if (!found) {
-          found = true;
-          seq = seq_pot;
-          *value = tmp;
-        } else if (seq_pot > seq) {
-          seq = seq_pot;
-          *value = tmp;
-        }
-      }
-    }
-    if (!found) {
-      s = current->Get(options, lkey, value);
-    }
-    mutex_.Lock();
-  }
-
-  // Ensures that old data can be removed.
-  for (size_t i = 0; i < ZnsConfig::lower_concurrency; i++) {
-    mem[i]->Unref();
-    if (imm[i] != nullptr) imm[i]->Unref();
-  }
-
-  current->Unref();
-  // printf("Gotten \n");
-
-  return s;
-}
-
-Status DBImplZNS::Get(const ReadOptions& options,
-                      ColumnFamilyHandle* column_family, const Slice& key,
-                      PinnableSlice* value, std::string* timestamp) {
-  std::string* val = new std::string;
-  Status s = Get(options, key, val);
-  *value = PinnableSlice(val);
-  return s;
-}
-
-Iterator* DBImplZNS::NewIterator(const ReadOptions& options,
-                                 ColumnFamilyHandle* column_family) {
-  return NULL;
-  // mutex_.Lock();
-  // SequenceNumber latest_snapshot = versions_->LastSequence();
-  // std::vector<Iterator*> list;
-  // // Memtables
-  // list.push_back(mem_->NewIterator());
-  // mem_->Ref();
-  // if (imm_ != nullptr) {
-  //   list.push_back(imm_->NewIterator());
-  //   imm_->Ref();
-  // }
-  // // Version
-  // versions_->current()->AddIterators(options, &list);
-  // versions_->current()->Ref();
-  // // Join iter
-  // Iterator* internal_iter =
-  //     NewMergingIterator(&internal_comparator_, &list[0], list.size());
-  // mutex_.Unlock();
-  // // Go to db iter
-  // return NewDBIterator(this, user_comparator(), internal_iter,
-  // latest_snapsshot,
-  //                      0);
 }
 
 int DBImplZNS::NumberLevels(ColumnFamilyHandle* column_family) {
